refactor: remove Wizard default exports and improve - fix resulting issues
