+++ src/components/Wizard/Presets/Loading.scss
@@ -1,0 +1,5 @@
+.zui-wizard-loading {
+  > p:first-of-type {
+    margin-top: 0;
+  }
+}
