+++ src/components/Wizard/Presets/index.ts
@@ -1,0 +1,5 @@
+export * from './Buttons';
+export * from './Confirmation';
+export * from './Header';
+export * from './Loading';
+export * from './NFTDetails';
