+++ src/components/Wizard/Wizard.scss
@@ -1,8 +1,3 @@
-@import '../../styles/typography';
-@import '../../styles/colors';
-
-$gutter: 16px;
-
 .zui-wizard {
   display: flex;
   flex-direction: column;
@@ -10,43 +5,4 @@ $gutter: 16px;
   box-sizing: border-box;
 
   gap: 2.5rem;
-
-  .zui-wizard-header {
-    width: 100%;
-    text-align: center;
-
-    * {
-      text-align: center;
-    }
-
-    h1 {
-      font-size: #{$size-large};
-      font-weight: #{$font-weight-bold};
-    }
-
-    h2 {
-      margin-top: 0;
-      font-size: 1rem;
-      font-weight: 400;
-    }
-
-    hr {
-      margin-top: 1.5rem;
-      margin-bottom: 0;
-      height: 1px;
-      background-color: #{$grey};
-      border: none;
-    }
-
-    .zui-wizard-header-container {
-      display: flex;
-      justify-content: center;
-    }
-  }
-
-  .zui-wizard-loading {
-    > p:first-of-type {
-      margin-top: 0;
-    }
-  }
 }
