+++ src/components/Wizard/Wizard.tsx
@@ -1,49 +1,22 @@
-import React, { FC, ReactNode } from 'react';
+import React, { FC } from 'react';
 
-// Style Imports
 import classNames from 'classnames';
 
-// Component Imports
-import { InfoTooltip } from '../InfoTooltip/InfoTooltip';
+import { Header, HeaderProps } from './Presets';
 
-// Preset Screens
-import { Buttons } from './Presets/Buttons';
-import { Confirmation } from './Presets/Confirmation';
-import { Loading } from './Presets/Loading';
 import './Wizard.scss';
-
-export interface HeaderProps {
-  header?: ReactNode;
-  headerInfo?: string;
-  subHeader?: string;
-  sectionDivider?: boolean;
-}
 
 export interface WizardProps extends HeaderProps {
   children: React.ReactNode;
   className?: string;
 }
 
-const Header: FC<HeaderProps> = ({ children, header, headerInfo, subHeader, sectionDivider = true }) => (
-  <div className={classNames('zui-wizard-header')}>
-    <div className="zui-wizard-header-container">
-      <h1>{header}</h1>
-      {headerInfo && <InfoTooltip content={headerInfo} />}
-    </div>
-    {subHeader && <h2>{subHeader}</h2>}
-    {children}
-    {sectionDivider && <hr />}
-  </div>
-);
-
-const Container: FC<WizardProps> = ({ children, className, ...headerProps }) => (
+export const Wizard: FC<WizardProps> = ({ children, className, header, headerInfo, subHeader, sectionDivider }) => (
   <div className={classNames('zui-wizard', className)}>
     {/* Header */}
-    {headerProps.header && <Header {...headerProps} />}
+    {header && <Header header={header} headerInfo={headerInfo} subHeader={subHeader} sectionDivider={sectionDivider} />}
 
     {/* Wizard Body */}
     {children}
   </div>
 );
-
-export const Wizard = { Container, Loading, Buttons, Confirmation, Header };
