+++ src/components/Wizard/index.ts
@@ -1,4 +1,1 @@
 export * from './Wizard';
-export type { ButtonsProps } from './Presets/Buttons';
-export type { ConfirmationProps } from './Presets/Confirmation';
-export type { LoadingProps } from './Presets/Loading';
